fix: AlarmChatPeopleBox 추가 - AlarmChatPeople에서 AlarmChatPeopleBox로 subscription 이동
