+++ front/src/UI/molecules/Alarm/AlarmChatPeople/AlarmChatPeopleQueries.ts
@@ -38,15 +38,3 @@ export const FORCED_OUT = gql`
     }
   }
 `;
-
-export const CHATLOG_SUBSCRIPTION = gql`
-  subscription onChatLogAdded($uuid: String!) {
-    chatLogAdded(uuid: $uuid) {
-      index
-      userID
-      message
-      type
-      createdAt
-    }
-  }
-`;
