+++ front/src/UI/organisms/AlarmChat/index.tsx
@@ -3,7 +3,6 @@ import { AccordionItem, AccordionButton, AccordionPanel, AccordionIcon, Box, Tex
 import { useQuery, useReactiveVar } from '@apollo/client';
 import { ContextMenu } from 'holee-contextmenu';
 
-import { AlarmChatPeople } from '../../molecules';
 import { AlarmChatMessagesBox } from '../AlarmChatMessagesBox';
 import { ChatLogSendBox } from '../ChatLogSendBox';
 import { PersonIcon, LockIcon } from '../../../utils/icons';
@@ -19,6 +18,7 @@ import {
 import { GET_CHAT, CHATLOG_SUBSCRIPTION } from './AlarmChatQueries';
 import { currentChatVar } from '../../../apollo/apolloProvider';
 import { EmptyChat } from '../../molecules/EmptyChat';
+import { AlarmChatPeopleBox } from '../AlarmChatPeopleBox';
 
 export const AlarmChat = () => {
   const [chatRoomState, setChatRoomState] = useState<string>('chat-room');
@@ -142,18 +142,7 @@ export const AlarmChat = () => {
             <ChatLogSendBox muteIDList={data.chat.muteID} />
           </>
         ) : (
-          <>
-            {data.chat.userID.map((username) => (
-              <AlarmChatPeople
-                key={username}
-                ownerID={data.chat.ownerID}
-                adminID={data.chat.adminID}
-                username={username}
-                refetchChat={refetch}
-                subscribeToNewMessage={subscribeToNewMessage}
-              />
-            ))}
-          </>
+          <AlarmChatPeopleBox chat={data.chat} refetchChat={refetch} subscribeToNewMessage={subscribeToNewMessage} />
         )}
       </AccordionPanel>
     </AccordionItem>
