+++ front/src/UI/organisms/AlarmChatPeopleBox/AlarmChatPeopleBoxQueries.ts
@@ -1,0 +1,13 @@
+import { gql } from '@apollo/client';
+
+export const CHATLOG_SUBSCRIPTION = gql`
+  subscription onChatLogAdded($uuid: String!) {
+    chatLogAdded(uuid: $uuid) {
+      index
+      userID
+      message
+      type
+      createdAt
+    }
+  }
+`;
