+++ front/src/UI/organisms/AlarmChatPeopleBox/index.tsx
@@ -1,0 +1,28 @@
+import React, { useEffect } from 'react';
+import { AlarmChatPeople } from '../../molecules';
+import { CHATLOG_SUBSCRIPTION } from './AlarmChatPeopleBoxQueries';
+
+export const AlarmChatPeopleBox = ({ ...props }) => {
+  const { chat, refetchChat, subscribeToNewMessage } = props;
+  useEffect(() => {
+    const unsubscribe = subscribeToNewMessage({
+      document: CHATLOG_SUBSCRIPTION,
+    });
+    return () => {
+      unsubscribe();
+    };
+  }, [subscribeToNewMessage]);
+  return (
+    <>
+      {chat.userID.map((username) => (
+        <AlarmChatPeople
+          key={username}
+          ownerID={chat.ownerID}
+          adminID={chat.adminID}
+          username={username}
+          refetchChat={refetchChat}
+        />
+      ))}
+    </>
+  );
+};
